requirements.txt
openai>=1.0.0
llama-index>=0.9.0
llama-index-readers-notion>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-vector-stores-chroma>=0.1.0
chromadb>=0.4.0
python-dotenv>=1.0.0
notion-client>=2.0.0
requests>=2.31.0
tqdm>=4.66.0
tabulate>=0.9.0
rich>=13.7.0